+++ SimLib/SimConnectForm.Position.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using Microsoft.FlightSimulator.SimConnect;
@@ -11,6 +12,8 @@ namespace SimLib
 {
     public partial class SimConnectForm : Form
     {
+        private TaskCompletionSource<Position> PositionTask;
+
         public void RegisterPosition()
         {
             try
@@ -21,7 +24,6 @@ namespace SimLib
                 simconnect.AddToDataDefinition(DEFINITIONS.Position, "Plane Altitude", "feet", SIMCONNECT_DATATYPE.FLOAT64, 0.0f, SimConnect.SIMCONNECT_UNUSED);
 
                 simconnect.RegisterDataDefineStruct<Position>(DEFINITIONS.Position);
-                simconnect.RequestDataOnSimObjectType(DATA_REQUESTS.Position, DEFINITIONS.Position, 0, SIMCONNECT_SIMOBJECT_TYPE.USER);
             }
             catch (COMException ex)
             {
@@ -29,38 +31,18 @@ namespace SimLib
             }
         }
 
+        public async Task<Position> GetPositionAsync()
+        {
+            PositionTask = new TaskCompletionSource<Position>();
+            simconnect.RequestDataOnSimObjectType(DATA_REQUESTS.Position, DEFINITIONS.Position, 0, SIMCONNECT_SIMOBJECT_TYPE.USER);
+
+            return await PositionTask.Task;
+        }
+
         private void OnRecvPosition(object sender, Position position)
         {
-            if (PositionChanged(position))
-                SimConnectPositionChanged(sender, new PositionChangedEventArgs() { position = position });
-
-            simconnect.RequestDataOnSimObjectType(DATA_REQUESTS.Position, DEFINITIONS.Position, 0, SIMCONNECT_SIMOBJECT_TYPE.USER);
+            if (PositionTask != null)
+                PositionTask.SetResult(position);
         }
-
-        private bool PositionChanged(Position position)
-        {
-            // if (position != LastPosition)
-                return true;
-        }
-
-        [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Ansi, Pack = 1)]
-        public class Position
-        {
-            // this is how you declare a fixed size string 
-            [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 256)]
-            public String title;
-            public double latitude;
-            public double longitude;
-            public double altitude;
-        }
-
-        public Position LastPosition;
-
-        public class PositionChangedEventArgs : EventArgs
-        {
-            public Position position
-            { get; internal set; }
-        }
-        public event EventHandler SimConnectPositionChanged;
     }
 }
