+++ SimLib/SimLib.csproj
@@ -45,6 +45,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Position.cs" />
     <Compile Include="SimConnectForm.Config.cs">
       <SubType>Form</SubType>
     </Compile>
